Developer

+++ app/src/main/java/com/example/varun/learningkotlinmvvm/Model/User.kt
@@ -6,7 +6,7 @@ import android.arch.persistence.room.PrimaryKey
 import java.util.*
 
 @Entity(tableName = "User")
-data class User(@PrimaryKey(autoGenerate = true) var id: Long?,
+data class User(@PrimaryKey(autoGenerate = true) var id: Int?,
                 @ColumnInfo(name = "name") var name: String,
                 @ColumnInfo(name = "email") var email: String
